Creating layout for 'MovieDetails' screen
Creating the layout for the page that shows the movie details when clicking on an item of the list shown on 'MovieScreen'.
We needed to change the constructor to get the data send from the 'MovieScreen' page. 

Also added a default image to the movieImages folder to start preparing a default value for the MovieDetails page.