Transfer onItemClick function
We transferred the 'onItemClick' function from MovieListItem to MovieScreen, this is a cleaner solution as it makes the MovieListItem component more generic.

We are also able to change what we transfer via MovieScreen, we send the 'onItemClick' instead of 'navigation'.
This prevents some security issues and also make it easier to change the function.